+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/command/config/ExampleUpdatedCallback.java
@@ -1,0 +1,5 @@
+package com.elmakers.mine.bukkit.magic.command.config;
+
+public interface ExampleUpdatedCallback {
+    void updated(boolean success, String exampleKey, String url);
+}